Rename remote logger directory
